Added Sun Position and Hillshade Unit Test
