+++ visibility/test/test_sun_position/TemplateConfigTest.py
@@ -24,77 +24,78 @@ import traceback
 
 import TestUtilities
 
-try:
-    print("Testing ArcPy")
-    arcpy.AddMessage("ArcPy works")
+def checkInputPaths():
+    try:
+        print("Testing ArcPy")
+        arcpy.AddMessage("ArcPy works")
 
-    # WORKAROUND: delete scratch db (having problems with scratch read-only "scheme lock" errors
-    # print "Deleting Scratch Workspace (Workaround)"    
-    # TestUtilities.deleteScratch()
+        # WORKAROUND: delete scratch db (having problems with scratch read-only "scheme lock" errors
+        # print "Deleting Scratch Workspace (Workaround)"    
+        # TestUtilities.deleteScratch()
 
-    if not arcpy.Exists(TestUtilities.scratchGDB):
-        TestUtilities.createScratch()
+        if not arcpy.Exists(TestUtilities.scratchGDB):
+            TestUtilities.createScratch()
 
-    print("Testing Necessary Paths")
+        print("Testing Necessary Paths")
 
-    print("Running from: " + str(TestUtilities.currentPath))
+        print("Running from: " + str(TestUtilities.currentPath))
 
-    paths2Check = []
-    paths2Check.extend([TestUtilities.geodatabasePath, TestUtilities.scratchPath, TestUtilities.toolboxesPath])
+        paths2Check = []
+        paths2Check.extend([TestUtilities.geodatabasePath, TestUtilities.scratchPath, TestUtilities.toolboxesPath])
 
-    for path2check in paths2Check :
-        if os.path.exists(path2check) :
-            print("Valid Path: " + path2check)
-        else :
-            print("ERROR: Necessary Path not found: " + path2check )
-            raise Exception('Bad Path')
+        for path2check in paths2Check :
+            if os.path.exists(path2check) :
+                print("Valid Path: " + path2check)
+            else :
+                print("ERROR: Necessary Path not found: " + path2check )
+                raise Exception('Bad Path')
 
-    # WORKAROUND
-    # print "Creating New Scratch Workspace (Workaround)"
-    # TestUtilities.createScratch()
+        # WORKAROUND
+        # print "Creating New Scratch Workspace (Workaround)"
+        # TestUtilities.createScratch()
 
-    print("Testing Necessary Geo Objects")
+        print("Testing Necessary Geo Objects")
 
-    objects2Check = []
-    objects2Check.extend([TestUtilities.toolbox, TestUtilities.inputGDB, TestUtilities.outputGDB,
+        objects2Check = []
+        objects2Check.extend([TestUtilities.toolbox, TestUtilities.inputGDB, TestUtilities.outputGDB,
                           TestUtilities.defaultGDB, TestUtilities.inputArea, TestUtilities.inputSurface,
                           TestUtilities.compareResults]) 
-    for object2Check in objects2Check :
-        desc = arcpy.Describe(object2Check)
-        if desc == None :
-            print("--> Invalid Object: " + str(object2Check)    )
-            arcpy.AddError("Bad Input")
-            raise Exception('Bad Input')
-        else :
-            print("Valid Object: " + desc.Name )
+        for object2Check in objects2Check :
+            desc = arcpy.Describe(object2Check)
+            if desc == None :
+                print("--> Invalid Object: " + str(object2Check)    )
+                arcpy.AddError("Bad Input")
+                raise Exception('Bad Input')
+            else :
+                print("Valid Object: " + desc.Name )
 
-    print("Test Successful")
+        print("Check Input Paths ran successfully.")
 
-except arcpy.ExecuteError: 
-    # Get the arcpy error messages
-    msgs = arcpy.GetMessages()
-    arcpy.AddError(msgs)
-    print(msgs)
+    except arcpy.ExecuteError: 
+        # Get the arcpy error messages
+        msgs = arcpy.GetMessages()
+        arcpy.AddError(msgs)
+        print(msgs)
 
-    # return a system error code
-    sys.exit(-1)
+        # return a system error code
+        sys.exit(-1)
 
-except:
-    # Get the traceback object
-    tb = sys.exc_info()[2]
-    tbinfo = traceback.format_tb(tb)[0]
+    except:
+        # Get the traceback object
+        tb = sys.exc_info()[2]
+        tbinfo = traceback.format_tb(tb)[0]
 
-    # Concatenate information together concerning the error into a message string
-    pymsg = "PYTHON ERRORS:\nTraceback info:\n" + tbinfo + "\nError Info:\n" + str(sys.exc_info()[1])
-    msgs = "ArcPy ERRORS:\n" + arcpy.GetMessages() + "\n"
+        # Concatenate information together concerning the error into a message string
+        pymsg = "PYTHON ERRORS:\nTraceback info:\n" + tbinfo + "\nError Info:\n" + str(sys.exc_info()[1])
+        msgs = "ArcPy ERRORS:\n" + arcpy.GetMessages() + "\n"
 
-    # Return python error messages for use in script tool or Python Window
-    arcpy.AddError(pymsg)
-    arcpy.AddError(msgs)
+        # Return python error messages for use in script tool or Python Window
+        arcpy.AddError(pymsg)
+        arcpy.AddError(msgs)
 
-    # Print Python error messages for use in Python / Python Window
-    print(pymsg + "\n")
-    print(msgs)
+        # Print Python error messages for use in Python / Python Window
+        print(pymsg + "\n")
+        print(msgs)
 
-    # return a system error code
-    sys.exit(-1)
+        # return a system error code
+        sys.exit(-1)
