+++ visibility/test/test_sun_position/TestSunPositionAndHillshade.py
@@ -25,71 +25,73 @@ import TestUtilities
 import os
 import datetime
 
-
-class ValueDifferenceError(Exception):
-    ''' throws errors if compared rasters are different '''
-    def __init__(self, message):
-        self.message = message
+import SunPositionAndHillshadeUnitTest
 
 try:
-    print("Importing toolbox... ")
-    arcpy.ImportToolbox(TestUtilities.toolbox, "sunpos")
-    arcpy.env.overwriteOutput = True
+    ###### Call Unit Test code
+    sunPosTestCase = SunPositionAndHillshadeUnitTest.SunPositionAndHillshadeUnitTest('test_sun_position_analysis')
+    sunPosTestCase.run()
+    
+    ######
+    
+    # print("Importing toolbox... ")
+    # arcpy.ImportToolbox(TestUtilities.toolbox, "sunpos")
+    # arcpy.env.overwriteOutput = True
 
-    # Inputs
-    print("Setting up inputs... ")
-    '''
-    Tool comparison is based on static dataset in Web Mercator
-    over Afghanistan, for 7/30/2015 at 14:28:36.
-    '''
-    dtCompare = datetime.datetime(2015, 7, 30, 14, 28, 36)
-    utcAfghanistan = r'(UTC+4:30) Afghanistan'
-    outHillshade = os.path.join(TestUtilities.outputGDB, "outputHS")
+    # # Inputs
+    # print("Setting up inputs... ")
+    # '''
+    # Tool comparison is based on static dataset in Web Mercator
+    # over Afghanistan, for 7/30/2015 at 14:28:36.
+    # '''
+    # dtCompare = datetime.datetime(2015, 7, 30, 14, 28, 36)
+    # utcAfghanistan = r'(UTC+4:30) Afghanistan'
+    # outHillshade = os.path.join(TestUtilities.outputGDB, "outputHS")
 
-    # Testing
-    arcpy.AddMessage("Running tool (Sun Position and Hillshade)")
-    arcpy.SunPositionAnalysis_sunpos(TestUtilities.inputArea, TestUtilities.inputSurface, dtCompare, utcAfghanistan, outHillshade)
+    # # Testing
+    # arcpy.AddMessage("Running tool (Sun Position and Hillshade)")
+    # arcpy.SunPositionAnalysis_sunpos(TestUtilities.inputArea, TestUtilities.inputSurface, dtCompare, utcAfghanistan, outHillshade)
 
-    print("Comparing expected values with tool results from " + str(dtCompare)\
-          + " in " + str(utcAfghanistan))
-    compareResults = TestUtilities.compareResults
+    # print("Comparing expected values with tool results from " + str(dtCompare)\
+          # + " in " + str(utcAfghanistan))
+    # compareResults = TestUtilities.compareResults
 
-    arcpy.CheckOutExtension("Spatial")
-    diff = Minus(Raster(outHillshade),Raster(compareResults))
-    diff.save(os.path.join(TestUtilities. scratchGDB, "diff"))
-    arcpy.CalculateStatistics_management(diff)
-    rasMinimum = float(arcpy.GetRasterProperties_management(diff,"MINIMUM").getOutput(0))
-    rasMaximum = float(arcpy.GetRasterProperties_management(diff,"MAXIMUM").getOutput(0))
-    rasMean = float(arcpy.GetRasterProperties_management(diff,"MEAN").getOutput(0))
-    rasSTD = float(arcpy.GetRasterProperties_management(diff,"STD").getOutput(0))
-    rasUnique = int(arcpy.GetRasterProperties_management(diff,"UNIQUEVALUECOUNT").getOutput(0))
+    # arcpy.CheckOutExtension("Spatial")
+    # diff = Minus(Raster(outHillshade),Raster(compareResults))
+    # diff.save(os.path.join(TestUtilities. scratchGDB, "diff"))
+    # arcpy.CalculateStatistics_management(diff)
+    # rasMinimum = float(arcpy.GetRasterProperties_management(diff,"MINIMUM").getOutput(0))
+    # rasMaximum = float(arcpy.GetRasterProperties_management(diff,"MAXIMUM").getOutput(0))
+    # rasMean = float(arcpy.GetRasterProperties_management(diff,"MEAN").getOutput(0))
+    # rasSTD = float(arcpy.GetRasterProperties_management(diff,"STD").getOutput(0))
+    # rasUnique = int(arcpy.GetRasterProperties_management(diff,"UNIQUEVALUECOUNT").getOutput(0))
 
-    if (rasMaximum == float(0)) and (rasMinimum == float(0)) and (rasUnique == int(1)):
-        print("No differences between tool output and expected results.")
-        print("Test Passed")
-    else:
-        msg = "ERROR IN TOOL RESULTS: \n"\
-            + "Difference between tool output and expected results found:\n"\
-            + "Difference Minimum: " + str(rasMinimum) + "\n"\
-            + "Difference Maximum: " + str(rasMaximum) + "\n"\
-            + "Difference Mean: " + str(rasMean) + "\n"\
-            + "Difference Std. Deviation: " + str(rasSTD) + "\n"\
-            + "Difference Number of Unique Values: " + str(rasUnique) + "\n"
-        raise ValueDifferenceError(msg)
+    # if (rasMaximum == float(0)) and (rasMinimum == float(0)) and (rasUnique == int(1)):
+        # print("No differences between tool output and expected results.")
+        # print("Test Passed")
+    # else:
+        # msg = "ERROR IN TOOL RESULTS: \n"\
+            # + "Difference between tool output and expected results found:\n"\
+            # + "Difference Minimum: " + str(rasMinimum) + "\n"\
+            # + "Difference Maximum: " + str(rasMaximum) + "\n"\
+            # + "Difference Mean: " + str(rasMean) + "\n"\
+            # + "Difference Std. Deviation: " + str(rasSTD) + "\n"\
+            # + "Difference Number of Unique Values: " + str(rasUnique) + "\n"
+        # raise ValueDifferenceError(msg)
 
 
-except arcpy.ExecuteError:
-    # Get the arcpy error messages
-    msgs = arcpy.GetMessages()
-    arcpy.AddError(msgs)
-    print(msgs)
+# except arcpy.ExecuteError:
+    # # Get the arcpy error messages
+    # msgs = arcpy.GetMessages()
+    # arcpy.AddError(msgs)
+    # print(msgs)
 
-    # return a system error code
-    sys.exit(-1)
+    # # return a system error code
+    # sys.exit(-1)
 
-except ValueDifferenceError as e:
-    print(e.message)
-    sys.exit(-1)
+# except ValueDifferenceError as e:
+    # print(e.message)
+    # sys.exit(-1)
 
 except:
     # Get the traceback object
@@ -112,5 +114,5 @@ except:
     # return a system error code
     sys.exit(-1)
 
-finally:
-    arcpy.CheckInExtension("Spatial")
+# finally:
+    # arcpy.CheckInExtension("Spatial")
