+++ visibility/test/test_sun_position/build.xml
@@ -10,17 +10,17 @@
   <property name="pythonpath" value="${dir.arcpy}\DLLs;${dir.arcpy}\lib;${dir.arcpy}\lib\plat-win;${dir.arcpy}\lib\lib-tk;${dir.arcpy};${dir.arcpy}\lib\site-packages;${dir.arcgisdesktop}\bin;${dir.arcgisdesktop}\arcpy"/>
 
   <target name="run_all_tests" 
-    depends="set_all_variables,test_config,test_sun_position">
+    depends="set_all_variables, test_sun_position">
     <echo message="All Tests Complete" />
   </target>
 
-  <target name="test_config" depends="set_all_variables">
+ <!--  <target name="test_config" depends="set_all_variables">
     <exec executable="${dir.arcpy}/python.exe" failonerror="true">
       <env key="PYTHONPATH" path="${pythonpath}"/>
       <arg line="-u"/>
       <arg line="${dir.src}/TemplateConfigTest.py"/>
     </exec>
-  </target>
+  </target> -->
   
   <target name="test_sun_position" depends="set_all_variables">
     <exec executable="${dir.arcpy}/python.exe" failonerror="true">
